Abandon the admin server.
This is a big shift. The Admin Server for 389 doesn't give me enough
value to justify the trouble I have with it. It doesn't work in secure
environments, and when you make the environment loose enough for the web
interface to actually be useful, it doesn't really give you anything
useful. So, LDAP server only, and no Admin Server.