Add a script that deletes outdated kernels.
I have a client that uses Oracle Linux and their Unbreakable
Enterprise kernel. As is normal, when new kernels get released,
quarterly, the hosts are updated. However, the old kernels are not
automatically deleted. In some cases, this has resulted in /boot
running out of space. So, prior to the monthly upgrade, run this script
to clean out the /boot filesystem in an RPM appropriate way.