Update the CentOS 8 kickstart notes
The `repo` directives won't follow URL schemes starting with `file://`.
So, the more I think about it, the more I'm inclined to delete the
BaseOS and AppStream directories from the modified ISO file. If the
installer isn't going to read/use them, why keep them? So, anticipate
another set of changes in the next week.

Signed-off-by: dafydd2277 <dafydd2277@gmail.com>

- modified:   kickstart/centos8/README.md
- modified:   kickstart/centos8/isoBuilder.sh
- modified:   kickstart/centos8/ks.cfg