Add ffmpeg to the package list
The ffmpeg libraries aren't installed by default in CentOS 8. This
means many video streaming options are unavailable. To make the
challenge even more challenging, the library isn't available in
any of the base repositories. So, let's add the library as part of
the initial installation, along with the repository it lives in.
For more information, see

https://forums.centos.org/viewtopic.php?t=74107

Signed-off-by: David Barr <dafydd@dafydd.com>

- modified:   kickstart/centos8/hostname.packages.cfg
- modified:   kickstart/centos8/repos.cfg

Signed-off-by: David Barr <dafydd@dafydd.com>